# swagger/docs/swagger.yaml
definitions:
...
  dao.TwoIndicator:
    properties:
      create_time:
        description: 创建时间
        type: string
      one_indicator_id:
        description: 所属一级指标
        type: integer
      score:
        description: 分值
        type: number
      two_indicator_id:
        description: 主键id
        type: integer
      two_indicator_name:
        description: 二级指标名称
        type: string
    type: object
...
  resp.CreateOrUpdateTwoIndicatorResp:
    properties:
      create_time:
        description: 创建时间
        type: string
      one_indicator_id:
        description: 所属一级指标
        type: integer
      score:
        description: 分值
        type: number
      two_indicator_id:
        description: 主键id
        type: integer
      two_indicator_name:
        description: 二级指标名称
        type: string
    type: object
...
  resp.GetOneIndicatorListResp:
    properties:
      list:
        items:
          $ref: '#/definitions/resp.OneIndicator'
        type: array
      total:
        type: integer
    type: object
...
  resp.GetTwoIndicatorListResp:
    properties:
      list:
        items:
          $ref: '#/definitions/resp.TwoIndicator'
        type: array
      total:
        type: integer
    type: object
...
  resp.OneIndicator:
    properties:
      content:
        description: 评分标准说明
        type: string
      count:
        description: 二级指标个数
        type: integer
      create_time:
        description: 创建时间
        type: string
      list:
        description: 二级指标列表
        items:
          $ref: '#/definitions/dao.TwoIndicator'
        type: array
      one_indicator_id:
        description: 主键id
        type: integer
      one_indicator_name:
        description: 一级指标名称
        type: string
    type: object
...
  resp.Response:
    properties:
      code:
        type: integer
      data: {}
      error_stack: {}
      msg:
        type: string
    type: object
...
  resp.TwoIndicator:
    properties:
      create_time:
        description: 创建时间
        type: string
      one_indicator_id:
        description: 所属一级指标
        type: integer
      one_indicator_name:
        description: 一级指标名称
        type: string
      score:
        description: 分值
        type: number
      two_indicator_id:
        description: 主键id
        type: integer
      two_indicator_name:
        description: 二级指标名称
        type: string
    type: object
...
host: localhost:7850
info:
  contact:
    email: support@swagger_controller.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a teachers-awards server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger teachers-awards API
  version: "1.0"
paths:
...
  /v1/export/awards/set/list:
    get:
      parameters:
      - description: 活动id
        in: query
        name: activity_id
        required: true
        type: integer
      - description: 用户姓名
        in: query
        name: user_name
        type: string
      - description: 1：男，2：女
        in: query
        name: user_sex
        type: integer
      - description: 科目code
        in: query
        name: subject_code
        type: string
      - description: 学校id
        in: query
        name: school_id
        type: string
      - description: 学校名称
        in: query
        name: school_name
        type: string
      - description: 申报类型
        in: query
        name: declare_type
        type: integer
      - description: 身份证号
        in: query
        name: identity_card
        type: string
      - description: 名次，0：无，1：一等奖，2：二等奖，3：三等奖
        in: query
        name: rank_prize
        type: integer
      - description: 最终得分（各项通过的审核）
        in: query
        name: final_score
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      summary: 导出奖项设置列表（教育局）
      tags:
      - Export
  /v1/export/edb/review/list:
    get:
      parameters:
      - description: 用户姓名
        in: query
        name: user_name
        type: string
      - description: 1：男，2：女
        in: query
        name: user_sex
        type: integer
      - description: 科目code
        in: query
        name: subject_code
        type: string
      - description: 学校id
        in: query
        name: school_id
        type: string
      - description: 学校名称
        in: query
        name: school_name
        type: string
      - description: 审核状态 0:全部 1:待审核，2：已审核
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      summary: 导出审核列表（教育局）
      tags:
      - Export
  /v1/export/history/activity/list:
    get:
      parameters:
      - description: 用户姓名
        in: query
        name: user_name
        type: string
      - description: 1：男，2：女
        in: query
        name: user_sex
        type: integer
      - description: 科目code
        in: query
        name: subject_code
        type: string
      - description: 学校id
        in: query
        name: school_id
        type: string
      - description: 学校名称
        in: query
        name: school_name
        type: string
      - description: 年份
        in: query
        name: year
        type: integer
      - description: 申报类型
        in: query
        name: declare_type
        type: integer
      - description: 身份证号
        in: query
        name: identity_card
        type: string
      - description: 0：无，1：一等奖，2：二等奖，3：三等奖
        in: query
        name: rank_prize
        type: integer
      - description: 排名
        in: query
        name: rank
        type: integer
      - description: 最终得分（各项通过的审核）
        in: query
        name: final_score
        type: number
      produces:
      - application/octet-stream
      responses: {}
      summary: 导出历史活动评审结果列表
      tags:
      - Export
  /v1/export/review/list:
    get:
      parameters:
      - description: 评委id
        in: query
        name: judges_id
        type: string
      - description: 用户姓名
        in: query
        name: user_name
        type: string
      - description: 1：男，2：女
        in: query
        name: user_sex
        type: integer
      - description: 科目code
        in: query
        name: subject_code
        type: string
      - description: 学校id
        in: query
        name: school_id
        type: string
      - description: 学校名称
        in: query
        name: school_name
        type: string
      - description: 年份
        in: query
        name: year
        type: integer
      - description: 一级指标名称
        in: query
        name: one_indicator_name
        type: string
      - description: 二级指标名称
        in: query
        name: two_indicator_name
        type: string
      - description: 审核状态 0:全部 1:待审核，2：未通过，3：已通过，4：已审核
        in: query
        name: review_status
        type: integer
      - description: 当前审核进程，1：学校，2：专家，3：教育局
        in: query
        name: review_process
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses: {}
      summary: 导出审核列表
      tags:
      - Export
  /v1/export/user/declare/record/list:
    get:
      parameters:
      - description: 年份
        in: query
        name: year
        required: true
        type: integer
      - description: 申报的用户id
        in: query
        name: declare_user_id
        required: true
        type: string
      - description: 角色类型，1：学校，2：专家，3：教育局，4：老师
        in: query
        name: role
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses: {}
      summary: 导出用户的申报记录
      tags:
      - Export
  /v1/indicator/one/delete/ids:
    delete:
      parameters:
      - description: 一级指标ids
        in: query
        name: one_indicator_ids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      summary: 删除一级指标
      tags:
      - Indicator
  /v1/indicator/one/list:
    get:
      parameters:
      - description: 输入的一级指标名称
        in: query
        name: input_name
        type: string
      - description: 输入的评分标准说明
        in: query
        name: input_content
        type: string
      - description: 是否获取二级指标
        in: query
        name: with_two
        type: boolean
      - description: 页数
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.GetOneIndicatorListResp'
              type: object
      summary: 获取一级指标列表
      tags:
      - Indicator
  /v1/indicator/one/save:
    post:
      parameters:
      - description: 一级指标id
        in: formData
        name: one_indicator_id
        type: integer
      - description: 一级指标名称
        in: formData
        name: one_indicator_name
        required: true
        type: string
      - description: 评分标准说明
        in: formData
        name: content
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      summary: 创建或者更新一级指标
      tags:
      - Indicator
  /v1/indicator/two/delete/ids:
    delete:
      parameters:
      - description: 二级指标ids
        in: query
        name: two_indicator_ids
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.Response'
      summary: 删除二级指标
      tags:
      - Indicator
  /v1/indicator/two/list:
    get:
      parameters:
      - description: 所属一级指标
        in: query
        name: one_indicator_id
        type: integer
      - description: 输入的二级指标名称
        in: query
        name: input_name
        type: string
      - description: 输入的分值
        in: query
        name: input_score
        type: number
      - description: 页数
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.GetTwoIndicatorListResp'
              type: object
      summary: 获取二级指标列表
      tags:
      - Indicator
  /v1/indicator/two/save:
    post:
      parameters:
      - description: 二级指标id
        in: formData
        name: two_indicator_id
        type: integer
      - description: 二级指标名称
        in: formData
        name: two_indicator_name
        required: true
        type: string
      - description: 分值
        in: formData
        name: score
        required: true
        type: number
      - description: 所属一级指标
        in: formData
        name: one_indicator_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/resp.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.CreateOrUpdateTwoIndicatorResp'
              type: object
      summary: 创建或者更新二级指标
      tags:
      - Indicator
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"